requirements.txt
beautifulsoup4==4.11.1
datasets==2.8.0
huggingface-hub==0.11.1
lxml==4.9.1
numpy==1.23.5
pandas==1.5.2
psutil==5.9.4
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.13.2
#torch==1.13.0
torch-tb-profiler==0.4.1
torchtext==0.14.0
torchvision==0.14.0
tqdm
transformers==4.27.4
pycocotools==2.0.6
openai==0.27.2
nltk==3.6.7
git+https://github.com/Maluuba/nlg-eval.git@master